+++ Source/Starfish.Service/UseCases/Team/TeamUpdateUseCase.cs
@@ -7,7 +7,7 @@ namespace Nerosoft.Starfish.UseCases;
 
 public interface ITeamUpdateUseCase : INonOutputUseCase<TeamUpdateInput>;
 
-public record TeamUpdateInput(int Id, TeamEditDto Data) : IUseCaseInput;
+public record TeamUpdateInput(long Id, TeamEditDto Data) : IUseCaseInput;
 
 public class TeamUpdateUseCase : ITeamUpdateUseCase
 {
