+++ Source/Starfish.Webapp/Rest/Defines/ITeamApi.cs
@@ -12,23 +12,23 @@ internal interface ITeamApi
 	Task<IApiResponse<int>> CountAsync([Query] TeamCriteria criteria, CancellationToken cancellationToken = default);
 
 	[Get("/api/team/{id}")]
-	Task<IApiResponse<TeamDetailDto>> GetAsync(int id, CancellationToken cancellationToken = default);
+	Task<IApiResponse<TeamDetailDto>> GetAsync(long id, CancellationToken cancellationToken = default);
 
 	[Post("/api/team")]
 	Task<IApiResponse> CreateAsync([Body] TeamEditDto data, CancellationToken cancellationToken = default);
 
 	[Put("/api/team/{id}")]
-	Task<IApiResponse> UpdateAsync(int id, [Body] TeamEditDto data, CancellationToken cancellationToken = default);
+	Task<IApiResponse> UpdateAsync(long id, [Body] TeamEditDto data, CancellationToken cancellationToken = default);
 
 	[Get("/api/team/{id}/member")]
-	Task<ApiResponse<List<TeamMemberDto>>> GetMembersAsync(int id, CancellationToken cancellationToken = default);
+	Task<ApiResponse<List<TeamMemberDto>>> GetMembersAsync(long id, CancellationToken cancellationToken = default);
 
 	[Post("/api/team/{id}/member")]
-	Task<IApiResponse> AppendMemberAsync(int id, [Body] List<int> userIds, CancellationToken cancellationToken = default);
+	Task<IApiResponse> AppendMemberAsync(long id, [Body] List<long> userIds, CancellationToken cancellationToken = default);
 
 	[Delete("/api/team/{id}/member")]
-	Task<IApiResponse> RemoveMemberAsync(int id, [Body] List<int> userIds, CancellationToken cancellationToken = default);
+	Task<IApiResponse> RemoveMemberAsync(long id, [Body] List<long> userIds, CancellationToken cancellationToken = default);
 
 	[Delete("/api/team/{id}/quit")]
-	Task<IApiResponse> QuitAsync(int id, CancellationToken cancellationToken = default);
+	Task<IApiResponse> QuitAsync(long id, CancellationToken cancellationToken = default);
 }
